pyproject.toml
[tool.poetry]
name = "dbt-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
pandas = "^2.2.0"
sqlalchemy = "^2.0.27"
python-dotenv = "^1.0.1"
dbt-postgres = "^1.7.8"
psycopg-binary = "^3.1.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
